msm: adsprpc: Support multiple sessions per process
Currently a process is limited to create only 2 sessions, by toggling
the 30th bit of tgid of the process, to create different process IDs on
DSP remote sybsystem. This approach is not scalable to create unique
process IDs to DSP, by using bits within the tgid of the process. Add
support to allow a process to create multiple sessions by choosing and
sending unique dsp process IDs on DSP remote sub system, instead of tgid
of HLOS process.

Change-Id: I33f52c68453301bdbb83dfb9a10df16143098a49
Acked-by: Santosh Sakore <quic_ssakore@quicinc.com>
Signed-off-by: Ansa Ahmed <quic_ansa@quicinc.com>