soc: qcom: ipcc: Add suspend-resume support
The value of register IPCC_REG_RECV_SIGNAL_ENABLE is not persistent,
so need to restore the unmask irq when triggered device wakeup.

The target platforms use PM_SUSPEND_TO_IDLE as its default
suspend target. Since syscore_ops's suspend/resume callbacks
are not called in this path, the IPCC driver will never be
able to log the client's details that caused the wakeup in
it's resume handler. Hence, move the suspend/resume handlers
to dev_pm_ops for better visibility.

Change-Id: I1fd4a2e5d72dd02714a4400f936e500e82fce016
Signed-off-by: Lijuan Gao <quic_lijuang@quicinc.com>
Signed-off-by: Shreyas K K <quic_shrekk@quicinc.com>