pci: msm: Correct link width setting to PCIE20_CAP and GEN2_CTRL_OFF
As per databook, when there are unused lanes in a system, you must
reprogram the following registers through the DBI. For this example, where
a 4-lane controller is connected to a 1-lane PHY:

1. Reprogram LINK_CAPABLE field of the PORT_LINK_CTRL_OFF register to 6h1
from 6h7. This is used by the LTSSM in Detect.

2. Reprogram NUM_OF_LANES[8:0] field of the GEN2_CTRL_OFF register to 9h1
from 9h4. This indicates to the LTSSM, the number of lanes to check for
exiting from L2.Idle or Polling.Active.

3. Reprogram PCIE_CAP_MAX_LINK_WIDTH field of the LINK_CAPABILITIES_REG
register to 6h1 from 6h4. This enables the RP to determine the Maximum
Link Width for this port.

We should not set link_width to PCIE20_CAP and GEN2_CTRL_OFF, but should
set the target_link_width. NUM_OF_LANES[8:0] field is recommended to set
as LINK_WIDTH_X1 even if there are more one lane.

Change-Id: I7b32160d6582b6ddc65d860b5e58f6ec77baf48d
Signed-off-by: Qiang Yu <quic_qianyu@quicinc.com>