drivers: arm-smmu: Power-on SMMU during freeze
There can be a possibility that runtime power management
has powered-down SMMU. Now, if there is a freeze request
on top of that, we will see SMMU register access errors,
as the SMMU is already powered-down.

Power-on SMMU during the freeze for accessing the SMMU
registers as part of qcom_free_io_pgtable_ops().

Change-Id: I6fff3ca0a70c071947e34bc4e574e3e565ee875d
Signed-off-by: Shreyas K K <quic_shrekk@quicinc.com>
Signed-off-by: Yadu MG <quic_ymg@quicinc.com>